Move shapes to types
