+++ src/utils/events.rs
@@ -3,7 +3,7 @@ use web_sys::{CanvasRenderingContext2d, HtmlCanvasElement};
 use yew::NodeRef;
 use yew::PointerEvent;
 
-use crate::store::shapes::Shape;
+use crate::types::shapes::Shape;
 use crate::utils::tools::{ToolAction, ToolBar};
 
 type Point = (f64, f64);
