+++ src/utils/tools/mod.rs
@@ -1,7 +1,7 @@
 use enum_dispatch::enum_dispatch;
 use web_sys::CanvasRenderingContext2d;
 
-use crate::store::shapes::{Ellipse, Rectangle, Shape};
+use crate::types::shapes::{Ellipse, Rectangle, Shape};
 use crate::utils::events::Event;
 pub mod select_tool;
 pub mod shape_tool;
